+++ src/main/java/br/devrafaelsoares/storeapirestful/exceptions/AuthStructureException.java
@@ -1,10 +1,10 @@
 package br.devrafaelsoares.storeapirestful.exceptions;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.jetbrains.annotations.NotNull;
 
 import java.time.ZonedDateTime;
