--- src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  application:
-    security:
-      jwt:
-        secret-key: 536173525767737261374d57736b45457436706358612f554768765a4d4738512b745267354f6747483162634939426d6e386b4f5056694f31614c306350534b51485a5735504a336163726e7244782f7541307a515a44674374552f557164623264457771686c71676b38336d5368462b77724f547433656441382f43614550464c79336a682f7250516b41534442356a4c55764b5738324d6a55674d51347a6850704a534653555345444e6c66705949676d2b465539583274424763523775476b5a59366a74484c6a4f2f484833446f4461446d684c63446f774d6c61642b6574724f597664757161474464636232624b4c5441302f77476f4e64726f6e78486e41722f62496e4b71447a5168506a5a4637746448446f76466663734e4459684953734559427050595a6f584e6d4855396b784e75784a4c6f3168344e757770474846474d316a44615a42474643636d6570533562644d4b58504b415779765135307a4a6958596d444d3da
-        expiration: 30
